Revert "We use the backup gem to back up the database and the /data/ folder."
This reverts commit 45d4c9d1f6b5db0e8446ef4649b872e97c6e0284.

That gem should just be installed on production servers for the backup job,
it's not an integral part or requirement of Tapestry.

Incidentally, including it in the Gemfile makes starting the drb server via
monit unhappy, because monit strips ENV[HOME] for security reasons, and the
backup gem does not take kindly to that.