The migration to rails3 introduced a bug where retaking an eligibility exam would not remove the previous answers. This commit fixes that.
It also relaxes the post only requirement for the retake url. A lot of people
hit that url with get instead of post - presumably when something goes wrong
with the jquery/javascript loading on their page.

This was tracked in bug #316 which has not been hit since the rails3 migration,
and which this commit closes.