Make the detection of the next enrollment step a big smarter.
New steps that are inserted will now show up as the next step, and be clickable
by unenrolled users.

This change will prompt unenrolled users who have taken the older version of
the eligibity questionnaire to take the current one, and will make those users
eligible for enrollment (if all the other requirements are met).

The 'Eligibility Questionnaire Results' step was added when we updated the
eligibility questionnaire. To reflect that this step may now be accessed by
people who have already taken the exam, this commit changes the link to the
next step to 'Continue', without referring to the name of the next step.